[codex] Unblock remaining PR checks
