+++ backend/internal/pipeline/api/hls_test.go
@@ -476,7 +476,7 @@ func TestServeHLS_StartingSegmentWaitsForArtifact(t *testing.T) {
 
 	assert.Equal(t, http.StatusOK, w.Code)
 	assert.Equal(t, "video/mp2t", w.Header().Get("Content-Type"))
-	assert.Equal(t, "public, max-age=60", w.Header().Get("Cache-Control"))
+	assert.Equal(t, "no-store", w.Header().Get("Cache-Control"))
 	assert.Equal(t, "identity", w.Header().Get("Content-Encoding"))
 	assert.Empty(t, w.Header().Get("X-XG2G-Reason"))
 	assert.Equal(t, "segment-data", w.Body.String())
